Remove AppFormSearch.vue
